test(agreement): introduce test fixtures
